Go Modules: /v2 suffix on module path
This was missed when initially migrating this project to Go Modules; it
was already at v2.x so should have had a version suffix added to the
module path. The docs [1] suggest incrementing the major version while
migrating from v2+ to Go Modules, however we have now already published
a v2.x tag with Go Modules, so we're going to stick with that.

[1]: https://github.com/golang/go/wiki/Modules#releasing-modules-v2-or-higher